refactor(db): chunk country_subdivision repository
This commit refactors the `country_subdivision` repository into the chunked pattern, breaking down the monolithic implementation into smaller, more manageable files. Both the standard and batch repository implementations have been updated, and all corresponding tests have been co-located with the code they validate.

Components affected:
- `banking-db-postgres`: Major refactoring of the `country_subdivision` repository.
- `.roo/commands`: Added a new, standardized `chunk-repo` command.
- `docs/progress`: Updated progress tracking documentation.

Benefits:
- **Improved Maintainability**: The chunked pattern makes the code easier to read, understand, and maintain.
- **Standardized Process**: The new `chunk-repo` command provides a consistent process for future repository refactoring.
- **Enhanced Testability**: Co-locating tests with their implementation simplifies the testing process.
