+++ banking-db-postgres/src/lib.rs
@@ -5,7 +5,7 @@ pub mod utils;
 pub use postgres_repositories::PostgresRepositories;
 pub use repository::audit_repository_impl::AuditLogRepositoryImpl;
 pub use repository::person::country_repository::repo_impl::CountryRepositoryImpl;
-pub use repository::person::country_subdivision_repository_impl::CountrySubdivisionRepositoryImpl;
+pub use repository::person::country_subdivision_repository::CountrySubdivisionRepositoryImpl;
 pub use repository::person::entity_reference_repository_impl::EntityReferenceRepositoryImpl;
 pub use repository::person::locality_repository_impl::LocalityRepositoryImpl;
 pub use repository::person::location_repository_impl::LocationRepositoryImpl;
