+++ banking-db-postgres/src/repository/person/country_subdivision_repository/exists_by_id.rs
@@ -1,0 +1,15 @@
+use banking_db::repository::CountrySubdivisionResult;
+use uuid::Uuid;
+
+use super::repo_impl::CountrySubdivisionRepositoryImpl;
+
+pub async fn exists_by_id(
+    repo: &CountrySubdivisionRepositoryImpl,
+    id: Uuid,
+) -> CountrySubdivisionResult<bool> {
+    Ok(repo
+        .country_subdivision_idx_cache
+        .read()
+        .await
+        .contains_primary(&id))
+}
