+++ banking-db-postgres/src/repository/person/locality_repository_impl.rs
@@ -6,7 +6,7 @@ use banking_db::repository::{
     TransactionAware,
 };
 use crate::repository::executor::Executor;
-use crate::repository::person::country_subdivision_repository_impl::CountrySubdivisionRepositoryImpl;
+use crate::repository::person::country_subdivision_repository::CountrySubdivisionRepositoryImpl;
 use crate::repository::person::location_repository_impl::LocationRepositoryImpl;
 use crate::utils::{get_heapless_string, get_optional_heapless_string, TryFromRow};
 use once_cell::sync::OnceCell;
