+++ banking-db-postgres/src/repository/person/mod.rs
@@ -1,6 +1,5 @@
 pub mod country_repository;
-pub mod country_subdivision_repository_impl;
-pub mod country_subdivision_repository_batch_impl;
+pub mod country_subdivision_repository;
 pub mod entity_reference_repository_batch_impl;
 pub mod entity_reference_repository_impl;
 pub mod locality_repository_impl;
@@ -9,3 +8,5 @@ pub mod location_repository_batch_impl;
 pub mod location_repository_impl;
 pub mod person_repository_batch_impl;
 pub mod person_repository_impl;
+#[cfg(test)]
+pub mod test_helpers;
