+++ docs/progress/progress-tracking.md
@@ -73,6 +73,11 @@ This commit significantly refactors the `CountryRepositoryImpl` by decomposing i
 - **Reduced File Size**: This refactoring drastically reduces the size of individual files, improving readability and LLM performance.
 - **Test Co-location**: Corresponding tests have been adjusted or co-located to align with the new modular structure.
 
+### 🎉 `CountrySubdivision` Repository Refactoring
+- **Chunked Pattern**: Refactored both standard and batch implementations into a unified `chunk-repo` pattern.
+- **Test Co-location**: Co-located tests with their corresponding method implementations for improved clarity.
+- **Standardized Command**: Created a new, standardized `chunk-repo` command to streamline future refactoring work.
+
 ## Key Achievements (September 2025)
 
 ### 🎉 EntityReference Batch Operations
